--- src/api/index.js
@@ -1,5 +1,0 @@
-import * as users from "./fake.api/user.api"
-const API = {
-    users
-}
-export default API
