+++ src/app/api/index.js
@@ -1,0 +1,5 @@
+import * as users from './fake.api/user.api';
+const API = {
+  users,
+};
+export default API;
